requirements.txt
numpy>=1.24
pillow>=10.0
scikit-image>=0.22
torch>=2.0
torchvision>=0.15
lpips>=0.1.4
plotly>=5.18
streamlit>=1.33
pandas>=2.0
orjson>=3.9
watchdog
